grouping of the unixsocket-related code
