+++ src/tcp.rs
@@ -1,26 +1,11 @@
 use std::net::ToSocketAddrs;
 
-use async_std::{
-    net::{TcpListener, TcpStream},
-    os::unix::net::{UnixListener, UnixStream},
-};
+use async_std::net::{TcpListener, TcpStream};
 
 use clap::Parser;
 use futures::{future::FutureExt, pin_mut, select};
 
 use crate::{std_socket_io, Args, Result, Socket};
-
-async fn run_unixstream_tasks(stream: UnixStream) -> Result<()> {
-    let socket = Socket::UnixSocketStream(stream.clone());
-    let socket2 = Socket::UnixSocketStream(stream);
-    let stdin_task = std_socket_io::stdin_to_stream(socket2).fuse();
-    let stdout_task = std_socket_io::socket_to_stdout(socket).fuse();
-    pin_mut!(stdin_task, stdout_task);
-    select! {
-        _res = stdin_task => _res,
-        _res = stdout_task => _res,
-    }
-}
 
 // // pub async fn run_server(bind_addr: &str, bind_port: Option<u16>) -> Result<()> {}
 
@@ -38,15 +23,6 @@ pub async fn run_tcp_server(bind_addr: &str, bind_port: u16) -> Result<()> {
             }
             run_tcpstream_tasks(&mut stream).await?
         }
-        Err(_) => {}
-    }
-    Ok(())
-}
-
-pub async fn run_unix_socket_server(fh: &str) -> Result<()> {
-    let listener = UnixListener::bind(fh).await?;
-    match listener.accept().await {
-        Ok((stream, _addr)) => run_unixstream_tasks(stream).await?,
         Err(_) => {}
     }
     Ok(())
